+++ e-shop/src/main/java/com/nguyenpham/oganicshop/service/ProductService.java
@@ -13,7 +13,7 @@ public interface ProductService {
     Product save(Product product);
     long countNumberProduct();
     List<ProductResponse> getAllProduct();
-    Page<Product> getProductsByCategory(long categoryId, int minPrice, int maxPrice, int pageNum, int pageSize, String sortField, String sortDir);
+    Page<Product> getProductsByCategory(long categoryId, double minPrice, double maxPrice, int pageNum, int pageSize, String sortField, String sortDir);
     Page<Product> getProductsByKeyword(String keyword, int pageNum, int pageSize, String sortField, String sortDir);
     int getAmountAvailable(String productUrl);
     boolean isProvideEnoughQuantity(String productUrl, int quantity);
